+++ frontend/src/components/LoginFormPage/index.js
@@ -1,60 +1,20 @@
 import React, { useState } from 'react';
-import * as sessionActions from '../../store/session';
-import { useDispatch, useSelector } from 'react-redux';
-import { Redirect } from 'react-router-dom';
+import { Modal } from '../../context/Modal';
+import LoginForm from './LoginForm';
 
-import './LoginForm.css';
-
-function LoginFormPage() {
-    const dispatch = useDispatch();
-    const sessionUser = useSelector(state => state.session.user);
-    const [credential, setCredential] = useState('');
-    const [password, setPassword] = useState('');
-    const [errors, setErrors] = useState([]);
-
-    if (sessionUser) return (
-        <Redirect to="/" />
-    );
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        setErrors([]);
-        return dispatch(sessionActions.login({ credential, password }))
-            .catch(async (res) => {
-                const data = await res.json();
-                if (data && data.errors) setErrors(data.errors);
-            });
-    }
+function LoginFormModal() {
+    const [showModal, setShowModal] = useState(false);
 
     return (
         <>
-            <div className='nav-organizer'></div>
-            <form onSubmit={handleSubmit}>
-                <ul>
-                    {errors.map((error, idx) => <li key={idx}>{error}</li>)}
-                </ul>
-                <label>
-                    Username or Email
-                    <input
-                        type="text"
-                        value={credential}
-                        onChange={(e) => setCredential(e.target.value)}
-                        required
-                    />
-                </label>
-                <label>
-                    Password
-                    <input
-                        type="password"
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                        required
-                    />
-                </label>
-                <button type="submit">Log In</button>
-            </form>
+            <button onClick={() => setShowModal(true)}>Log In</button>
+            {showModal && (
+                <Modal onClose={() => setShowModal(false)}>
+                    <LoginForm />
+                </Modal>
+            )}
         </>
     );
 }
 
-export default LoginFormPage;
+export default LoginFormModal;
